requirements.txt
annotated-types==0.7.0
arrow==1.4.0
asgiref==3.10.0
async-timeout==5.0.1
bcrypt==5.0.0
binaryornot==0.4.4
camel-converter==5.0.0
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
cookiecutter==2.6.0
croniter==6.0.0
cryptography==46.0.3
distlib==0.4.0
Django==5.2.7
django-browser-reload==1.21.0
django-cloudinary-storage==0.3.0
django-ninja==1.4.5
django-rq==2.10.2
django-storages==1.14.6
django-tailwind==4.4.0
djangorestframework==3.16.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
honcho==2.0.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
meilisearch==0.37.1
msgpack==1.1.2
passlib==1.7.4
pillow==12.0.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
requests==2.32.5
rich==14.2.0
rq==2.6.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
